+++ packages/medusa/src/interfaces/index.ts
@@ -1,4 +1,6 @@
 export * from "./tax-calculation-strategy"
 export * from "./cart-completion-strategy"
 export * from "./tax-service"
-export * from "./base-service"
+export * from "./transaction-base-service"
+export * from "./models/base-entity"
+export * from "./models/soft-deletable-entity"