+++ packages/medusa/src/models/product.ts
@@ -2,8 +2,6 @@ import _ from "lodash"
 import {
   BeforeInsert,
   Column,
-  CreateDateColumn,
-  DeleteDateColumn,
   Entity,
   Index,
   JoinColumn,
@@ -11,11 +9,8 @@ import {
   ManyToMany,
   ManyToOne,
   OneToMany,
-  PrimaryColumn,
-  UpdateDateColumn,
 } from "typeorm"
-import { ulid } from "ulid"
-import { DbAwareColumn, resolveDbType } from "../utils/db-aware-column"
+import { DbAwareColumn } from "../utils/db-aware-column"
 import { Image } from "./image"
 import { ProductCollection } from "./product-collection"
 import { ProductOption } from "./product-option"
@@ -23,6 +18,8 @@ import { ProductTag } from "./product-tag"
 import { ProductType } from "./product-type"
 import { ProductVariant } from "./product-variant"
 import { ShippingProfile } from "./shipping-profile"
+import { SoftDeletableEntity } from "../interfaces/models/soft-deletable-entity"
+import { generateEntityId } from "../utils/generate-entity-id"
 
 export enum Status {
   DRAFT = "draft",
@@ -32,10 +29,7 @@ export enum Status {
 }
 
 @Entity()
-export class Product {
-  @PrimaryColumn()
-  id: string
-
+export class Product extends SoftDeletableEntity {
   @Column()
   title: string
 
@@ -72,17 +66,12 @@ export class Product {
   @Column({ nullable: true })
   thumbnail: string
 
-  @OneToMany(
-    () => ProductOption,
-    (productOption) => productOption.product
-  )
+  @OneToMany(() => ProductOption, (productOption) => productOption.product)
   options: ProductOption[]
 
-  @OneToMany(
-    () => ProductVariant,
-    (variant) => variant.product,
-    { cascade: true }
-  )
+  @OneToMany(() => ProductVariant, (variant) => variant.product, {
+    cascade: true,
+  })
   variants: ProductVariant[]
 
   @Index()
@@ -151,24 +140,14 @@ export class Product {
   @Column({ nullable: true })
   external_id: string
 
-  @CreateDateColumn({ type: resolveDbType("timestamptz") })
-  created_at: Date
-
-  @UpdateDateColumn({ type: resolveDbType("timestamptz") })
-  updated_at: Date
-
-  @DeleteDateColumn({ type: resolveDbType("timestamptz") })
-  deleted_at: Date
-
   @DbAwareColumn({ type: "jsonb", nullable: true })
-  metadata: any
+  metadata: Record<string, unknown>
 
   @BeforeInsert()
-  private beforeInsert() {
+  private beforeInsert(): void {
     if (this.id) return
-    const id = ulid()
-    this.id = `prod_${id}`
 
+    this.id = generateEntityId(this.id, "prod")
     if (!this.handle) {
       this.handle = _.kebabCase(this.title)
     }